+++ src/itg3200.h
@@ -36,7 +36,7 @@
 
 class ITG3200 {
     uint8_t buffer[6];   // Buffer for general use. Increase size as needed.
-    uint16_t gRaw[3];   // Raw bits received from ITG-3200
+    int16_t gRaw[3];   // Raw bits received from ITG-3200
     float gVec[3];   // [-2000, 2000] deg/s mapped to [-1, 1].
     bool calibrated;   // Disable integration until calibration finishes.
     float tempData[3];   // Temporary storage of calibration data
@@ -46,7 +46,7 @@ class ITG3200 {
 
     #ifdef GYRO_LPF_DEPTH
     int lpfIndex;
-    float lpfVal[3][GYRO_LPF_DEPTH];   // Low-pass filter values for each of three axes
+    int16_t lpfVal[3][GYRO_LPF_DEPTH];   // Low-pass filter values for each of three axes
     #endif // GYRO_LPF_DEPTH
